Add inlines.min.js to fix issues in prod (#124)
Both Django and django-material-admin use a minified version of a few JS files, one of which is inlines.js. But django-material-admin overrides it, which means that when `DEBUG=False`, the non-minified version gets loaded and everything is fine. In prod, with `DEBUG=True`, the minified version loads, which is from Django, and causes all kinds of issues.